requirements.txt
Django==4.2.8
djangorestframework==3.15.3
django-cors-headers==4.4.0
celery==5.6.0
redis==5.0.0
python-dotenv==1.0.1
django-environ==0.11.2
